requirements.txt
asyncio
aiohttp
requests
pillow
pyyaml